Align structure of header section
Following after the previous changes, we had to align the website
sign up text and the images. I aligned the various elements with
one another so that the structure of the header is in place and
all that it is missing is styling of its various elements. They
are now in the correct positions and are also a bit flexible in
order to adjust to some screen variance. Now the links are at
the top, and the second section of the header is at the bottom
and they are spaced sensibly between one another.