Add testimonial author and position it
Positioned the testimonial author, also a couple of minor other ch-
anges I made were to decrease the size of the testimonial and the
random information section slightly, as I realised they were sligh-
tly too big compared to the header. The rest of the changes were to
the styling to make the testimonial section look somewhat nice.